requirements-prod.txt
# Production dependencies only
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
bitsandbytes>=0.42.0
rouge-score
qwen-vl-utils
accelerate
Pillow
pymupdf
numpy

# API
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
jsonschema>=4.20.0
psycopg2-binary>=2.9.0
